Simplify boolean check
The result of the equality check can be used here.
This makes the logic a bit clearer.

Extract contains check

This is the underlying algorithm of this function.
Pulling it out makes `validArgs` a bit clearer what it is doing.

Generalize the type

Use the `inSet` function that already exists

Should have looked for this first :).

Let the language do the casting

I am not sure if this is an improvement, but it cleans up `validArgs`.